requirements.txt
asgiref==3.7.2
boto3==1.29.5
botocore==1.32.5
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-auto-prefetching==0.2.12
django-cors-headers==4.3.1
django-filter==23.4
django-rest-passwordreset==1.3.0
django-ses==3.5.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
idna==3.4
inflection==0.5.1
jmespath==1.0.1
packaging==23.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.18
